storage: rename files "bdev" -> "storage"
