+++ src/lxc/cgroups/cgfsng.c
@@ -50,11 +50,11 @@
 #include <linux/types.h>
 #include <linux/kdev_t.h>
 
-#include "bdev.h"
 #include "cgroup.h"
 #include "cgroup_utils.h"
 #include "commands.h"
 #include "log.h"
+#include "storage.h"
 #include "utils.h"
 
 lxc_log_define(lxc_cgfsng, lxc);
