+++ src/lxc/start.c
@@ -63,7 +63,6 @@
 #endif
 
 #include "af_unix.h"
-#include "bdev.h"
 #include "caps.h"
 #include "cgroup.h"
 #include "commands.h"
@@ -80,6 +79,7 @@
 #include "monitor.h"
 #include "namespace.h"
 #include "start.h"
+#include "storage.h"
 #include "storage_utils.h"
 #include "sync.h"
 #include "utils.h"
