+++ src/lxc/storage/loop.c
@@ -34,9 +34,9 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 
-#include "bdev.h"
 #include "log.h"
-#include "lxcloop.h"
+#include "loop.h"
+#include "storage.h"
 #include "storage_utils.h"
 #include "utils.h"
