+++ src/lxc/storage/loop.h
