+++ src/lxc/storage/lvm.c
@@ -32,11 +32,11 @@
 #include <sys/sysmacros.h>
 #include <sys/wait.h>
 
-#include "bdev.h"
 #include "config.h"
 #include "log.h"
-#include "lxclvm.h"
-#include "lxcrsync.h"
+#include "lvm.h"
+#include "rsync.h"
+#include "storage.h"
 #include "storage_utils.h"
 #include "utils.h"
