+++ src/lxc/storage/nbd.c
@@ -31,9 +31,9 @@
 #include <sys/prctl.h>
 #include <sys/wait.h>
 
-#include "bdev.h"
 #include "log.h"
-#include "lxcnbd.h"
+#include "nbd.h"
+#include "storage.h"
 #include "storage_utils.h"
 #include "utils.h"
