+++ src/lxc/storage/storage_utils.h
